+++ app/src/main/java/com/concordium/wallet/data/model/SimpleFraction.kt
@@ -1,0 +1,16 @@
+package com.concordium.wallet.data.model
+
+import java.math.BigInteger
+
+/**
+ * A fraction defined as 2 integers:
+ * ```
+ *  N
+ *  –
+ *  D
+ * ```
+ */
+data class SimpleFraction(
+    val numerator: BigInteger,
+    val denominator: BigInteger,
+)